+++ templates/admin-panel/template/client/__tests__/unit/components/pagination/pages.js
@@ -1,0 +1,10 @@
+import * as directives from 'sham-ui-directives';
+import PaginationPages  from '../../../../src/components/pagination/pages.sfc';
+import renderer from 'sham-ui-test-helpers';
+
+it( 'renders correctly', () => {
+    const meta = renderer( PaginationPages, {
+        directives
+    } );
+    expect( meta.toJSON() ).toMatchSnapshot();
+} );